add new book(ready) and update existing book

+++ src/main/java/com/chern/libraryapp/service/book/BookServiceImpl.java
@@ -4,6 +4,7 @@ import com.chern.libraryapp.dao.DAOFactory;
 import com.chern.libraryapp.model.Author;
 import com.chern.libraryapp.model.Book;
 import com.chern.libraryapp.model.Genre;
+import com.chern.libraryapp.model.enums.BookStatus;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -33,7 +34,7 @@ public class BookServiceImpl implements BookService {
 
     @Override
     public void updateBook(Book book, List<Author> newAuthors, List<Genre> newGenres) {
-//        DAOFactory.bookDAO().updateBook(book);
+        DAOFactory.bookDAO().updateBook(book);
         List<Genre> oldBookGenres = DAOFactory.genreDao().getBookGenresById(book.getId());
         List<Genre> oldBookGenresClone = new ArrayList<>(oldBookGenres);
         List<Genre> newBookGenresClone = new ArrayList<>(newGenres);
@@ -43,9 +44,31 @@ public class BookServiceImpl implements BookService {
                 newGenres.remove(newGenre);
             }
         }
-        // new insert old delete
-        DAOFactory.genreDao().addNewGenresToBook(newGenres, book.getId());
-        DAOFactory.genreDao().deleteSeveralGenresFromBook(oldBookGenres, book.getId());
-//        List<Author> oldBookAuthors = DAOFactory.authorDAO().getBookAuthorsById(book.getId());
+        if (newGenres != null){
+            DAOFactory.genreDao().addNewGenresToBook(newGenres, book.getId());
+        }
+        if (oldBookGenres != null){
+            DAOFactory.genreDao().deleteSeveralGenresFromBook(oldBookGenres, book.getId());
+        }
+
+        DAOFactory.authorDAO().addNewAuthors(newAuthors); // добавление в бд новых авторов
+        List<Author> newAuthorsWithId = DAOFactory.authorDAO().getNewAuthorsWithId(newAuthors);
+        DAOFactory.authorDAO().getBookAuthorsById(book.getId());
+        DAOFactory.authorDAO().deleteOldAuthors(book.getId());
+        DAOFactory.authorDAO().addNewAuthorsToBook(newAuthorsWithId, book.getId());
+
+    }
+
+    @Override
+    public void addNewBook(Book book) {
+        if (book.getTotalAmount() > 0)
+            book.setStatus(BookStatus.AVAILABLE);
+        else book.setStatus(BookStatus.UNAVAILABLE);
+        DAOFactory.bookDAO().addNewBook(book);
+        Book bookByISBN = DAOFactory.bookDAO().findBookByISBN(book.getIsbn()); // получаем книгу с ид
+        DAOFactory.authorDAO().addNewAuthors(book.getAuthors()); // заносим авторов в бд
+        List<Author> newAuthorsWithId = DAOFactory.authorDAO().getNewAuthorsWithId(book.getAuthors()); // получаем новых авторов книги с ид
+        DAOFactory.authorDAO().addNewAuthorsToBook(newAuthorsWithId, bookByISBN.getId()); // добавляем авторов для книги
+        DAOFactory.genreDao().addNewGenresToBook(book.getGenres(), bookByISBN.getId()); // добавляем жанры для книги
     }
 }
